bug-detector/requirements.txt
claude-agent-sdk>=0.1.4
groq>=0.4.0
python-dotenv>=1.0.0
click>=8.3.0
